% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomColor.R
\name{randomColor}
\alias{randomColor}
\title{Random colors}
\usage{
randomColor(n, hue = "random", luminosity = "random")
}
\arguments{
\item{n}{number of colors to be generated}

\item{hue}{the hue xxx}

\item{luminosity}{xxx}
}
\value{
A character vector of hexadecimal color codes.
}
\description{
Generate random colors.
}
\examples{
n <- 20
clrs <- randomColor(n, hue = "red", luminosity = "bright")
opar <- par(mar=c(0, 0, 0, 0))
pie(rep(1, n), col = clrs)
par(opar)
}
